backend/product/views.py
from django.shortcuts import render
from rest_framework import viewsets, views
from .models import Product
from .serializers import ProductSerializer

class ProductViewSet(viewsets.ModelViewSet):
    serializer_class = ProductSerializer
    queryset = Product.objects.all()
